Working on gui mvc app
